requirements.txt
autogen-agentchat
autogen-ext[azure]
python-dotenv
azure-identity